+++ kem/src/avx2/kyber_poly_avx2.c
@@ -30,10 +30,6 @@
 #include "kyber_poly_avx2.h"
 #include "shake_4x_avx2.h"
 
-#if (LC_KYBER_K != 4)
-#error "AVX2 support for Kyber mode 4 only"
-#endif
-
 void poly_compress(uint8_t r[LC_KYBER_POLYCOMPRESSEDBYTES],
 		   const poly *restrict a);
 
@@ -48,6 +44,7 @@ void poly_compress(uint8_t r[LC_KYBER_POLYCOMPRESSEDBYTES],
  *	    LC_KYBER_POLYCOMPRESSEDBYTES)
  * @param a pointer to input polynomial
  */
+#if (LC_KYBER_POLYCOMPRESSEDBYTES == 160)
 void poly_compress_avx(uint8_t r[LC_KYBER_POLYCOMPRESSEDBYTES],
 		       const poly *restrict a)
 {
@@ -154,6 +151,86 @@ void poly_decompress_avx(poly *restrict r,
 	}
 	LC_FPU_DISABLE;
 }
+
+#elif (LC_KYBER_POLYCOMPRESSEDBYTES == 128)
+void poly_compress_avx(uint8_t r[LC_KYBER_POLYCOMPRESSEDBYTES],
+		       const poly * restrict a)
+{
+	unsigned int i;
+	__m256i f0, f1, f2, f3;
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wdeclaration-after-statement"
+	const __m256i v = _mm256_load_si256(&kyber_qdata.vec[_16XV / 16]);
+	const __m256i shift1 = _mm256_set1_epi16(1 << 9);
+	const __m256i mask = _mm256_set1_epi16(15);
+	const __m256i shift2 = _mm256_set1_epi16((16 << 8) + 1);
+	const __m256i permdidx = _mm256_set_epi32(7, 3 ,6, 2, 5, 1, 4, 0);
+#pragma GCC diagnostic pop
+
+	LC_FPU_ENABLE;
+	for (i = 0; i < LC_KYBER_N / 64; i++) {
+		f0 = _mm256_load_si256(&a->vec[4 * i + 0]);
+		f1 = _mm256_load_si256(&a->vec[4 * i + 1]);
+		f2 = _mm256_load_si256(&a->vec[4 * i + 2]);
+		f3 = _mm256_load_si256(&a->vec[4 * i + 3]);
+		f0 = _mm256_mulhi_epi16(f0, v);
+		f1 = _mm256_mulhi_epi16(f1, v);
+		f2 = _mm256_mulhi_epi16(f2, v);
+		f3 = _mm256_mulhi_epi16(f3, v);
+		f0 = _mm256_mulhrs_epi16(f0, shift1);
+		f1 = _mm256_mulhrs_epi16(f1, shift1);
+		f2 = _mm256_mulhrs_epi16(f2, shift1);
+		f3 = _mm256_mulhrs_epi16(f3, shift1);
+		f0 = _mm256_and_si256(f0, mask);
+		f1 = _mm256_and_si256(f1, mask);
+		f2 = _mm256_and_si256(f2, mask);
+		f3 = _mm256_and_si256(f3, mask);
+		f0 = _mm256_packus_epi16(f0, f1);
+		f2 = _mm256_packus_epi16(f2, f3);
+		f0 = _mm256_maddubs_epi16(f0, shift2);
+		f2 = _mm256_maddubs_epi16(f2, shift2);
+		f0 = _mm256_packus_epi16(f0, f2);
+		f0 = _mm256_permutevar8x32_epi32(f0, permdidx);
+		_mm256_storeu_si256((__m256i_u *)&r[32 * i], f0);
+	}
+	LC_FPU_DISABLE;
+}
+
+void poly_decompress_avx(poly * restrict r,
+			 const uint8_t a[LC_KYBER_POLYCOMPRESSEDBYTES])
+{
+	unsigned int i;
+	__m128i t;
+	__m256i f;
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wdeclaration-after-statement"
+	const __m256i q = _mm256_load_si256(&kyber_qdata.vec[_16XQ/16]);
+	const __m256i shufbidx = _mm256_set_epi8(7, 7, 7, 7, 6, 6, 6, 6, 5, 5,
+						 5, 5, 4, 4, 4, 4, 3, 3, 3, 3,
+						 2, 2, 2, 2, 1, 1, 1, 1, 0, 0,
+						 0, 0);
+	const __m256i mask = _mm256_set1_epi32(0x00F0000F);
+	const __m256i shift = _mm256_set1_epi32((128 << 16) + 2048);
+#pragma GCC diagnostic pop
+
+	LC_FPU_ENABLE;
+	for (i =0; i < LC_KYBER_N / 16; i++) {
+		t = _mm_loadl_epi64((__m128i_u *)&a[8 * i]);
+		f = _mm256_broadcastsi128_si256(t);
+		f = _mm256_shuffle_epi8(f, shufbidx);
+		f = _mm256_and_si256(f, mask);
+		f = _mm256_mullo_epi16(f, shift);
+		f = _mm256_mulhrs_epi16(f, q);
+		_mm256_store_si256(&r->vec[i], f);
+	}
+	LC_FPU_DISABLE;
+}
+
+#else
+#error "Kyber AVX2 support incomplete"
+#endif
 
 /**
  * @brief poly_frommsg